Start building the app
There are a few modules so far. They should be self-explanatory... or at
least enough for the time being. Everything is using mock/dummy data.
Data persistence doesn't work yet. The timeline is the main
consolidated view of all user activity in the app. It's a reverse
chronological feed of feeds, incorporating notes (text, images, etc.)
that the user has logged themselves in addition to snippets of memories
from the past. The Rewind function is a place for viewing fun
recollections of things that happened recently - in effect a Spotify
Wrapped for your everyday life. The journals feature is an entrypoint
nto some really cool stuff that's being worked out, but rest assured:
it's a place for collecting and aggregating user memories in a fun
way. The library tab is a collection of collections of content
collected throughout the app (yes, I know how vague that is) - on
other words, the library is the place to search and reference all
types of media in a regularized format. This is also the place where
a user can see what data of theirs is backed up.

TODO: Implement data persistence