Begin implementing account authentication module
The :core:account module is a library for managing the account and
identity of the current user for LogDate. Due to the one-person,
one-account identity model of LogDate, we will assume that regardless of
operating system, the user on the current profile is the only one who
will be accessing the app; only one account can be added at a time. In
the future, LogDate will support generic service providers, allowing the
app to authenticate with any service that fulfills the account spec,
similar to client apps for Mastodon.

The :feature:account module will support UI and UI hooks to actually
allow a user to sign in and sign up.

TODO: Allow sign-up and sign-in to LogDate Cloud
TODO: Figure out how to use passkeys
TODO: Please for the love of everything right in the universe remove the
commented code in AccountCreationScreen.kt when able
TODO: Make :core:account module OS-agnostic